requirements.txt
fastapi
uvicorn
openai
pydantic
sqlalchemy
psycopg2-binary
pandas
matplotlib
python-dotenv
requests